+++ src/pine/ast/count.clj
@@ -1,0 +1,6 @@
+(ns pine.ast.count)
+
+;; todo: spec for the :value for a :limit
+
+(defn handle [acc value]
+  (assoc acc :count value))